add ScatterChart
